[exporter] Implement file splitting
Since the previous attempt (#95) caused component definition corruption, extract several classes into separate files without modifying the component class itself.
