+++ Editor/AssetPathUtils.cs
@@ -1,0 +1,52 @@
+// SPDX-FileCopyrightText: 2024-present hkrn
+// SPDX-License-Identifier: MPL
+
+using System;
+using System.IO;
+using UnityEditor;
+using UnityEngine;
+
+// ReSharper disable once CheckNamespace
+namespace com.github.hkrn
+{
+    internal static class AssetPathUtils
+    {
+        private const string CloneSuffix = "(Clone)";
+
+        public static string BasePath => "Assets/NDMF VRM Exporter";
+
+        public static string GetOutputPath(GameObject gameObject)
+        {
+            return GetBasePath(gameObject, BasePath);
+        }
+
+        public static string GetTempPath(GameObject gameObject)
+        {
+            return GetBasePath(gameObject, FileUtil.GetUniqueTempPathInProject());
+        }
+
+        public static string TrimCloneSuffix(string name)
+        {
+            while (name.EndsWith(CloneSuffix, StringComparison.Ordinal))
+            {
+                name = name[..^CloneSuffix.Length];
+            }
+
+            return name;
+        }
+
+        private static string GetBasePath(GameObject gameObject, string basePath)
+        {
+            var sceneName = !string.IsNullOrEmpty(gameObject.scene.name)
+                ? StripInvalidFileNameCharacters(gameObject.scene.name)
+                : "Untitled";
+            var gameObjectName = TrimCloneSuffix(StripInvalidFileNameCharacters(gameObject.name));
+            return $"{basePath}/{sceneName}/{gameObjectName}";
+        }
+
+        private static string StripInvalidFileNameCharacters(string name)
+        {
+            return string.Join("_", name.Split(Path.GetInvalidFileNameChars()));
+        }
+    }
+}
