+++ Editor/AssetPathUtils.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: adcb52077c9b48ada0518c544e7ce7f1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
