+++ Editor/MaterialBaker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b6ee1920c680446db7427e3ef48ead1a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
