+++ Editor/NDMFVRMExporter.cs
@@ -10,7 +10,6 @@ using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Net;
-using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
@@ -36,23 +35,9 @@ using VRC.SDKBase;
 using VRC.SDKBase.Editor.Api;
 #endif // NVE_HAS_VRCHAT_AVATAR_SDK
 
-#if NVE_HAS_AVATAR_OPTIMIZER
-using Anatawa12.AvatarOptimizer.API;
-using nadena.dev.ndmf.platform;
-#endif // NVE_HAS_AVATAR_OPTIMIZER
-
-#if NVE_HAS_LILYCAL_INVENTORY
-using jp.lilxyzw.lilycalinventory.runtime;
-#endif // NVE_HAS_LILYCAL_INVENTORY
-
 #if NVE_HAS_LILTOON
-using lilToon;
 using UnityEngine.Rendering;
 #endif // NVE_HAS_LILTOON
-
-#if NVE_HAS_MODULAR_AVATAR
-using nadena.dev.modular_avatar.core;
-#endif // NVE_HAS_MODULAR_AVATAR
 
 #if NVE_HAS_NDMF
 using nadena.dev.ndmf;
@@ -1147,149 +1132,6 @@ namespace com.github.hkrn
         public MaterialVariantMapping[] Mappings { get; init; } = { };
     }
 
-    internal static class UnityExtensions
-    {
-        public static System.Numerics.Vector2 ToVector2(this Vector2 self)
-        {
-            return new System.Numerics.Vector2(self.x, self.y);
-        }
-
-        public static System.Numerics.Vector2 ToVector2WithCoordinateSpace(this Vector2 self)
-        {
-            return new System.Numerics.Vector2(self.x, 1.0f - self.y);
-        }
-
-        public static System.Numerics.Vector3 ToVector3(this Vector3 self)
-        {
-            return new System.Numerics.Vector3(self.x, self.y, self.z);
-        }
-
-        public static System.Numerics.Vector3 ToVector3WithCoordinateSpace(this Vector3 self)
-        {
-            return new System.Numerics.Vector3(-self.x, self.y, self.z);
-        }
-
-        public static System.Numerics.Vector4 ToVector4WithTangentSpace(this Vector4 self)
-        {
-            var v = (new Vector3(-self.x, self.y, self.z)).normalized;
-            var w = Mathf.Sign(-self.w);
-            return new System.Numerics.Vector4(v.x, v.y, v.z, w);
-        }
-
-        public static System.Numerics.Vector3 ToVector3(this Color self)
-        {
-            return new System.Numerics.Vector3(self.r, self.g, self.b);
-        }
-
-        public static System.Numerics.Vector3 ToVector3(this Color self, ColorSpace src, ColorSpace dst)
-        {
-            var color = ConvertColor(self, src, dst);
-            return new System.Numerics.Vector3(color.r, color.g, color.b);
-        }
-
-        public static System.Numerics.Vector4 ToVector4(this Color self, ColorSpace src, ColorSpace dst)
-        {
-            var color = ConvertColor(self, src, dst);
-            return new System.Numerics.Vector4(color.r, color.g, color.b, color.a);
-        }
-
-        public static System.Numerics.Quaternion ToQuaternionWithCoordinateSpace(this Quaternion self)
-        {
-            self.ToAngleAxis(out var angle, out var axis);
-            var newAxis = axis.ToVector3WithCoordinateSpace();
-            var rotation = System.Numerics.Quaternion.CreateFromAxisAngle(newAxis, -angle * Mathf.Deg2Rad);
-            return System.Numerics.Quaternion.Normalize(rotation);
-        }
-
-        private static System.Numerics.Matrix4x4 ToMatrix4(this Matrix4x4 self)
-        {
-            return new System.Numerics.Matrix4x4(
-                self.m00, self.m01, self.m02, self.m03,
-                self.m10, self.m11, self.m12, self.m13,
-                self.m20, self.m21, self.m22, self.m23,
-                self.m30, self.m31, self.m32, self.m33);
-        }
-
-        public static System.Numerics.Matrix4x4 ToNormalizedMatrix(this Matrix4x4 self)
-        {
-            var t = new Vector3(-self.m03, self.m13, self.m23);
-            self.rotation.ToAngleAxis(out var angle, out var axis);
-            var r = Quaternion.AngleAxis(-angle, new Vector3(-axis.x, axis.y, axis.z)).normalized;
-            var s = new Vector3
-            {
-                x = new Vector4(self.m00, self.m10, self.m20, self.m30).magnitude,
-                y = new Vector4(self.m01, self.m11, self.m21, self.m31).magnitude,
-                z = new Vector4(self.m02, self.m12, self.m22, self.m32).magnitude,
-            };
-            return Matrix4x4.TRS(t, r, s).ToMatrix4();
-        }
-
-        public static gltf.material.TextureFilterMode ToTextureFilterMode(this FilterMode self)
-        {
-            return self switch
-            {
-                FilterMode.Point => gltf.material.TextureFilterMode.Nearest,
-                FilterMode.Bilinear => gltf.material.TextureFilterMode.Linear,
-                FilterMode.Trilinear => gltf.material.TextureFilterMode.LinearMipmapLinear,
-                _ => throw new ArgumentOutOfRangeException(),
-            };
-        }
-
-        public static gltf.material.TextureWrapMode ToTextureWrapMode(this TextureWrapMode self)
-        {
-            return self switch
-            {
-                TextureWrapMode.Clamp => gltf.material.TextureWrapMode.ClampToEdge,
-                TextureWrapMode.Mirror => gltf.material.TextureWrapMode.MirroredRepeat,
-                TextureWrapMode.Repeat => gltf.material.TextureWrapMode.Repeat,
-                _ => throw new ArgumentOutOfRangeException(),
-            };
-        }
-
-        internal static Texture2D Blit(this Texture sourceTexture, TextureFormat textureFormat, ColorSpace cs,
-            Material? material = null)
-        {
-            return Blit(sourceTexture, sourceTexture.width, sourceTexture.height, textureFormat, cs, material);
-        }
-
-        internal static Texture2D Blit(this Texture sourceTexture, int width, int height, TextureFormat textureFormat,
-            ColorSpace cs, Material? material)
-        {
-            var rw = cs == ColorSpace.Gamma ? RenderTextureReadWrite.sRGB : RenderTextureReadWrite.Linear;
-            var activeRenderTexture = RenderTexture.active;
-            var destRenderTexture = RenderTexture.GetTemporary(width, height, 0, RenderTextureFormat.Default, rw);
-            if (material)
-            {
-                Graphics.Blit(sourceTexture, destRenderTexture, material);
-            }
-            else
-            {
-                Graphics.Blit(sourceTexture, destRenderTexture);
-            }
-
-            RenderTexture.active = destRenderTexture;
-            var outputTexture = new Texture2D(width, height, textureFormat, false, cs == ColorSpace.Linear)
-            {
-                name = sourceTexture.name
-            };
-            outputTexture.ReadPixels(new Rect(0, 0, width, height), 0, 0);
-            RenderTexture.active = activeRenderTexture;
-            RenderTexture.ReleaseTemporary(destRenderTexture);
-            return outputTexture;
-        }
-
-        private static Color ConvertColor(Color self, ColorSpace src, ColorSpace dst)
-        {
-            var color = (src, dst) switch
-            {
-                (ColorSpace.Gamma, ColorSpace.Linear) => self.linear,
-                (ColorSpace.Linear, ColorSpace.Gamma) => self.gamma,
-                _ => self
-            };
-            return color;
-        }
-    }
-
     internal static class Translator
     {
         internal static readonly Localizer Instance = new("en-us", () =>
@@ -1300,470 +1142,6 @@ namespace com.github.hkrn
         });
 
         public static string _(string key) => Instance.GetLocalizedString(key);
-    }
-
-#if NVE_HAS_LILTOON
-    // Based on lilMaterialBaker
-    // https://github.com/lilxyzw/lilToon/blob/f80e1f13f385c1774a6626cb585a4ab6f5728caa/Assets/lilToon/Editor/lilMaterialBaker.cs
-    internal static class MaterialBaker
-    {
-        public static Texture? AutoBakeMainTexture(IAssetSaver assetSaver, Material material)
-        {
-            var mainColor = material.GetColor(PropertyColor);
-            var mainTexHsvg = material.GetVector(PropertyMainTexHsvg);
-            var mainGradationStrength = material.GetFloat(PropertyMainGradationStrength);
-            var useMainSecondTex = material.GetFloat(PropertyUseMainSecondTex);
-            var useMainThirdTex = material.GetFloat(PropertyUseMainThirdTex);
-            var mainTex = material.GetTexture(PropertyMainTex);
-            var shouldNotBakeAll = mainColor == Color.white && mainTexHsvg == lilConstants.defaultHSVG &&
-                                   mainGradationStrength == 0.0 && useMainSecondTex == 0.0 && useMainThirdTex == 0.0;
-            if (shouldNotBakeAll)
-                return null;
-            var mainSecondUVMode = material.HasFloat(PropertyMainSecondUVMode)
-                ? material.GetFloat(PropertyMainSecondUVMode)
-                : 0.0;
-            var mainThirdUVMode = material.HasFloat(PropertyMainThirdUVMode)
-                ? material.GetFloat(PropertyMainThirdUVMode)
-                : 0.0;
-            var bakeSecond = useMainSecondTex != 0.0 && mainSecondUVMode == 0.0;
-            var bakeThird = useMainThirdTex != 0.0 && mainThirdUVMode == 0.0;
-            // run bake
-            var hsvgMaterial = new Material(lilShaderManager.ltsbaker);
-
-            var mainTexName = mainTex ? mainTex!.name : AssetPathUtils.TrimCloneSuffix(material.name);
-            var srcTexture = new Texture2D(2, 2);
-            var srcMain2 = new Texture2D(2, 2);
-            var srcMain3 = new Texture2D(2, 2);
-            var srcMask2 = new Texture2D(2, 2);
-            var srcMask3 = new Texture2D(2, 2);
-
-            hsvgMaterial.SetColor(PropertyColor, Color.white);
-            hsvgMaterial.SetVector(PropertyMainTexHsvg, mainTexHsvg);
-            hsvgMaterial.SetFloat(PropertyMainGradationStrength, mainGradationStrength);
-            CopyTexture(hsvgMaterial, material, PropertyMainGradationTex);
-            CopyTexture(hsvgMaterial, material, PropertyMainColorAdjustMask);
-
-            AssignMaterialTexture(assetSaver, mainTex, hsvgMaterial, PropertyMainTex, ref srcTexture);
-            if (bakeSecond)
-            {
-                CopyFloat(hsvgMaterial, material, PropertyUseMainSecondTex);
-                CopyColor(hsvgMaterial, material, PropertyMainColorSecond);
-                CopyFloat(hsvgMaterial, material, PropertyMainSecondTexAngle);
-                CopyFloat(hsvgMaterial, material, PropertyMainSecondTexIsDecal);
-                CopyFloat(hsvgMaterial, material, PropertyMainSecondTexIsLeftOnly);
-                CopyFloat(hsvgMaterial, material, PropertyMainSecondTexIsRightOnly);
-                CopyFloat(hsvgMaterial, material, PropertyMainSecondTexShouldCopy);
-                CopyFloat(hsvgMaterial, material, PropertyMainSecondTexShouldFlipMirror);
-                CopyFloat(hsvgMaterial, material, PropertyMainSecondTexShouldFlipCopy);
-                CopyFloat(hsvgMaterial, material, PropertyMainSecondTexIsMsdf);
-                CopyFloat(hsvgMaterial, material, PropertyMainSecondTexBlendMode);
-                CopyFloat(hsvgMaterial, material, PropertyMainSecondTexAlphaMode);
-                CopyTextureOffset(hsvgMaterial, material, PropertyMainSecondTex);
-                CopyTextureScale(hsvgMaterial, material, PropertyMainSecondTex);
-                CopyTextureOffset(hsvgMaterial, material, PropertyMainSecondBlendMask);
-                CopyTextureScale(hsvgMaterial, material, PropertyMainSecondBlendMask);
-
-                AssignMaterialTexture(assetSaver, material.GetTexture(PropertyMainSecondTex), hsvgMaterial,
-                    PropertyMainSecondTex, ref srcMain2);
-                AssignMaterialTexture(assetSaver, material.GetTexture(PropertyMainSecondBlendMask), hsvgMaterial,
-                    PropertyMainSecondBlendMask, ref srcMask2);
-            }
-
-            if (bakeThird)
-            {
-                CopyFloat(hsvgMaterial, material, PropertyUseMainThirdTex);
-                CopyColor(hsvgMaterial, material, PropertyMainColorThird);
-                CopyFloat(hsvgMaterial, material, PropertyMainThirdTexAngle);
-                CopyFloat(hsvgMaterial, material, PropertyMainThirdTexIsDecal);
-                CopyFloat(hsvgMaterial, material, PropertyMainThirdTexIsLeftOnly);
-                CopyFloat(hsvgMaterial, material, PropertyMainThirdTexIsRightOnly);
-                CopyFloat(hsvgMaterial, material, PropertyMainThirdTexShouldCopy);
-                CopyFloat(hsvgMaterial, material, PropertyMainThirdTexShouldFlipMirror);
-                CopyFloat(hsvgMaterial, material, PropertyMainThirdTexShouldFlipCopy);
-                CopyFloat(hsvgMaterial, material, PropertyMainThirdTexIsMsdf);
-                CopyFloat(hsvgMaterial, material, PropertyMainThirdTexBlendMode);
-                CopyFloat(hsvgMaterial, material, PropertyMainThirdTexAlphaMode);
-                CopyTextureOffset(hsvgMaterial, material, PropertyMainThirdTex);
-                CopyTextureScale(hsvgMaterial, material, PropertyMainThirdTex);
-                CopyTextureOffset(hsvgMaterial, material, PropertyMainThirdBlendMask);
-                CopyTextureScale(hsvgMaterial, material, PropertyMainThirdBlendMask);
-
-                AssignMaterialTexture(assetSaver, material.GetTexture(PropertyMainThirdTex), hsvgMaterial,
-                    PropertyMainThirdTex, ref srcMain3);
-                AssignMaterialTexture(assetSaver, material.GetTexture(PropertyMainThirdBlendMask), hsvgMaterial,
-                    PropertyMainThirdBlendMask, ref srcMask3);
-            }
-
-            var outTexture = RunBake(srcTexture, hsvgMaterial);
-            outTexture.name = $"{mainTexName}_{nameof(AutoBakeMainTexture)}";
-            assetSaver.SaveAsset(outTexture);
-
-            Object.DestroyImmediate(hsvgMaterial);
-            Object.DestroyImmediate(srcTexture);
-            Object.DestroyImmediate(srcMain2);
-            Object.DestroyImmediate(srcMain3);
-            Object.DestroyImmediate(srcMask2);
-            Object.DestroyImmediate(srcMask3);
-
-            return outTexture;
-        }
-
-        public static Texture? AutoBakeShadowTexture(IAssetSaver assetSaver, Material material,
-            Texture? baseMainTexture, int shadowType = 0)
-        {
-            var useShadow = material.GetFloat(PropertyUseShadow);
-            var shadowColor = material.GetColor(PropertyShadowColor);
-            var shadowColorTex = material.GetTexture(PropertyShadowColorTex);
-            var shadowStrength = material.GetFloat(PropertyShadowStrength);
-            var shadowStrengthMask = material.GetTexture(PropertyShadowStrengthMask);
-            var mainTex = material.GetTexture(PropertyMainTex);
-
-            var shouldNotBakeAll = useShadow == 0.0 && shadowColor == Color.white && !shadowColorTex &&
-                                   !shadowStrengthMask;
-            if (shouldNotBakeAll)
-                return null;
-            // run bake
-            var hsvgMaterial = new Material(lilShaderManager.ltsbaker);
-
-            var mainTexName = mainTex ? mainTex!.name : AssetPathUtils.TrimCloneSuffix(material.name);
-            var srcTexture = new Texture2D(2, 2);
-            var srcMain2 = new Texture2D(2, 2);
-            var srcMask2 = new Texture2D(2, 2);
-
-            hsvgMaterial.SetColor(PropertyColor, Color.white);
-            hsvgMaterial.SetVector(PropertyMainTexHsvg, lilConstants.defaultHSVG);
-            hsvgMaterial.SetFloat(PropertyUseMainSecondTex, 1.0f);
-            hsvgMaterial.SetFloat(PropertyUseMainThirdTex, 1.0f);
-            hsvgMaterial.SetColor(PropertyMainColorThird,
-                new Color(1.0f, 1.0f, 1.0f, material.GetFloat(PropertyShadowMainStrength)));
-            hsvgMaterial.SetFloat(PropertyMainThirdTexBlendMode, 3.0f);
-
-            Texture? shadowTex;
-            switch (shadowType)
-            {
-                case 2:
-                {
-                    var shadowSecondColor = material.GetColor(PropertyShadowSecondColor);
-                    hsvgMaterial.SetColor(PropertyMainColorSecond,
-                        new Color(shadowSecondColor.r, shadowSecondColor.g, shadowSecondColor.b,
-                            shadowSecondColor.a * shadowStrength));
-                    hsvgMaterial.SetFloat(PropertyMainSecondTexBlendMode, 0.0f);
-                    hsvgMaterial.SetFloat(PropertyMainSecondTexAlphaMode, 0.0f);
-                    shadowTex = material.GetTexture(PropertyShadowSecondColorTex);
-                    break;
-                }
-                case 3:
-                {
-                    var shadowThirdColor = material.GetColor(PropertyShadowThirdColor);
-                    hsvgMaterial.SetColor(PropertyMainColorThird,
-                        new Color(shadowThirdColor.r, shadowThirdColor.g, shadowThirdColor.b,
-                            shadowThirdColor.a * shadowStrength));
-                    hsvgMaterial.SetFloat(PropertyMainThirdTexBlendMode, 0.0f);
-                    hsvgMaterial.SetFloat(PropertyMainThirdTexAlphaMode, 0.0f);
-                    shadowTex = material.GetTexture(PropertyShadowThirdColorTex);
-                    break;
-                }
-                default:
-                    hsvgMaterial.SetColor(PropertyMainColorSecond,
-                        new Color(shadowColor.r, shadowColor.g, shadowColor.b, shadowStrength));
-                    hsvgMaterial.SetFloat(PropertyMainSecondTexBlendMode, 0.0f);
-                    hsvgMaterial.SetFloat(PropertyMainSecondTexAlphaMode, 0.0f);
-                    shadowTex = material.GetTexture(PropertyShadowColorTex);
-                    break;
-            }
-
-            var referenceMainTexture =
-                LoadTexture(assetSaver, baseMainTexture, ref srcTexture) ? srcTexture : Texture2D.whiteTexture;
-            var referenceMainSecondTexture =
-                LoadTexture(assetSaver, shadowTex, ref srcMain2) ? srcMain2 : referenceMainTexture;
-            hsvgMaterial.SetTexture(PropertyMainTex, referenceMainTexture);
-            hsvgMaterial.SetTexture(PropertyMainSecondTex, referenceMainSecondTexture);
-            hsvgMaterial.SetTexture(PropertyMainThirdTex, referenceMainTexture);
-
-            var referenceMaskTexture =
-                LoadTexture(assetSaver, shadowStrengthMask, ref srcMask2) ? srcMask2 : Texture2D.whiteTexture;
-            hsvgMaterial.SetTexture(PropertyMainSecondBlendMask, referenceMaskTexture);
-            hsvgMaterial.SetTexture(PropertyMainThirdBlendMask, referenceMaskTexture);
-
-            var outTexture = RunBake(srcTexture, hsvgMaterial);
-            outTexture.name = $"{mainTexName}_{nameof(AutoBakeShadowTexture)}";
-            assetSaver.SaveAsset(outTexture);
-
-            Object.DestroyImmediate(hsvgMaterial);
-            Object.DestroyImmediate(srcTexture);
-            Object.DestroyImmediate(srcMain2);
-            Object.DestroyImmediate(srcMask2);
-
-            return outTexture;
-        }
-
-        public static Texture? AutoBakeMatCap(IAssetSaver assetSaver, Material material)
-        {
-            var matcapColor = material.GetColor(PropertyMatCapColor);
-            var matcapTex = material.GetTexture(PropertyMatCapTex);
-            var shouldNotBakeAll = matcapColor == Color.white;
-            if (shouldNotBakeAll)
-                return matcapTex;
-            // run bake
-            var bufMainTexture = matcapTex as Texture2D;
-            var hsvgMaterial = new Material(lilShaderManager.ltsbaker);
-
-            var matcapTexName = matcapTex ? matcapTex!.name : AssetPathUtils.TrimCloneSuffix(material.name);
-            var srcTexture = new Texture2D(2, 2);
-
-            hsvgMaterial.SetColor(PropertyColor, matcapColor);
-            hsvgMaterial.SetVector(PropertyMainTexHsvg, lilConstants.defaultHSVG);
-            AssignMaterialTexture(assetSaver, bufMainTexture, hsvgMaterial, PropertyMainTex, ref srcTexture);
-
-            var outTexture = RunBake(srcTexture, hsvgMaterial);
-            outTexture.name = $"{matcapTexName}_{nameof(AutoBakeMatCap)}";
-            assetSaver.SaveAsset(outTexture);
-
-            Object.DestroyImmediate(hsvgMaterial);
-            Object.DestroyImmediate(srcTexture);
-
-            return outTexture;
-        }
-
-        public static Texture? AutoBakeAlphaMask(IAssetSaver assetSaver, Material material)
-        {
-            var mainTex = material.GetTexture(PropertyMainTex);
-            // run bake
-            var bufMainTexture = mainTex as Texture2D;
-            var hsvgMaterial = new Material(lilShaderManager.ltsbaker);
-
-            var mainTexName = mainTex ? mainTex!.name : AssetPathUtils.TrimCloneSuffix(material.name);
-            var srcTexture = new Texture2D(2, 2);
-            var srcAlphaMask = new Texture2D(2, 2);
-
-            hsvgMaterial.EnableKeyword("_ALPHAMASK");
-            hsvgMaterial.SetColor(PropertyColor, Color.white);
-            hsvgMaterial.SetVector(PropertyMainTexHsvg, lilConstants.defaultHSVG);
-            CopyFloat(hsvgMaterial, material, PropertyAlphaMaskMode);
-            CopyFloat(hsvgMaterial, material, PropertyAlphaMaskScale);
-            CopyFloat(hsvgMaterial, material, PropertyAlphaMaskValue);
-
-            var baseTex = material.GetTexture(PropertyAlphaMask);
-            Texture2D? outTexture = null;
-            if (LoadTexture(assetSaver, baseTex, ref srcAlphaMask))
-            {
-                hsvgMaterial.SetTexture(PropertyAlphaMask, srcAlphaMask);
-                AssignMaterialTexture(assetSaver, bufMainTexture, hsvgMaterial, PropertyMainTex, ref srcTexture);
-                outTexture = RunBake(srcTexture, hsvgMaterial);
-                outTexture.name = $"{mainTexName}_{nameof(AutoBakeAlphaMask)}";
-                assetSaver.SaveAsset(outTexture);
-            }
-
-            Object.DestroyImmediate(hsvgMaterial);
-            Object.DestroyImmediate(srcTexture);
-
-            return outTexture;
-        }
-
-        private static Texture2D RunBake(Texture2D srcTexture, Material material,
-            Texture2D? referenceTexture = null)
-        {
-            int width = 4096, height = 4096;
-            if (referenceTexture)
-            {
-                width = referenceTexture!.width;
-                height = referenceTexture.height;
-            }
-            else if (srcTexture)
-            {
-                width = srcTexture.width;
-                height = srcTexture.height;
-            }
-
-            var outTexture = new Texture2D(width, height);
-            var bufRT = RenderTexture.active;
-            var dstTexture = RenderTexture.GetTemporary(width, height);
-            Graphics.Blit(srcTexture, dstTexture, material);
-            RenderTexture.active = dstTexture;
-            outTexture.ReadPixels(new Rect(0, 0, width, height), 0, 0);
-            RenderTexture.active = bufRT;
-            RenderTexture.ReleaseTemporary(dstTexture);
-            return outTexture;
-        }
-
-        private static void AssignMaterialTexture(IAssetSaver assetSaver, Texture? baseTex, Material material,
-            int propertyID, ref Texture2D srcTexture)
-        {
-            var result = LoadTexture(assetSaver, baseTex, ref srcTexture);
-            material.SetTexture(propertyID, result ? srcTexture : Texture2D.whiteTexture);
-        }
-
-        private static bool LoadTexture(IAssetSaver assetSaver, Texture? baseTexture, ref Texture2D srcTexture)
-        {
-            if (baseTexture is Texture2D { isReadable: true } readableTexture)
-            {
-                srcTexture.LoadImage(readableTexture.EncodeToPNG());
-            }
-            else if (baseTexture && assetSaver.IsTemporaryAsset(baseTexture))
-            {
-                Object.Destroy(srcTexture);
-                var innerBaseTexture = baseTexture!;
-                srcTexture = new Texture2D(innerBaseTexture.width, innerBaseTexture.height);
-                Graphics.ConvertTexture(innerBaseTexture, srcTexture);
-            }
-            else
-            {
-                var path = AssetDatabase.GetAssetPath(baseTexture);
-                if (string.IsNullOrEmpty(path))
-                {
-                    return false;
-                }
-
-                lilTextureUtils.LoadTexture(ref srcTexture, path);
-            }
-
-            return true;
-        }
-
-        private static void CopyFloat(Material dest, Material source, int name)
-        {
-            if (source.HasProperty(name))
-            {
-                dest.SetFloat(name, source.GetFloat(name));
-            }
-        }
-
-        private static void CopyColor(Material dest, Material source, int name)
-        {
-            if (source.HasProperty(name))
-            {
-                dest.SetColor(name, source.GetColor(name));
-            }
-        }
-
-        private static void CopyTexture(Material dest, Material source, int name)
-        {
-            if (source.HasProperty(name))
-            {
-                dest.SetTexture(name, source.GetTexture(name));
-            }
-        }
-
-        private static void CopyTextureOffset(Material dest, Material source, int name)
-        {
-            if (source.HasProperty(name))
-            {
-                dest.SetTextureOffset(name, source.GetTextureOffset(name));
-            }
-        }
-
-        private static void CopyTextureScale(Material dest, Material source, int name)
-        {
-            if (source.HasProperty(name))
-            {
-                dest.SetTextureScale(name, source.GetTextureScale(name));
-            }
-        }
-
-        private static readonly int PropertyColor = Shader.PropertyToID("_Color");
-        private static readonly int PropertyMainTexHsvg = Shader.PropertyToID("_MainTexHSVG");
-        private static readonly int PropertyMainGradationStrength = Shader.PropertyToID("_MainGradationStrength");
-        private static readonly int PropertyMainTex = Shader.PropertyToID("_MainTex");
-        private static readonly int PropertyMainSecondTex = Shader.PropertyToID("_Main2ndTex");
-        private static readonly int PropertyMainSecondBlendMask = Shader.PropertyToID("_Main2ndBlendMask");
-        private static readonly int PropertyMainThirdTex = Shader.PropertyToID("_Main3rdTex");
-        private static readonly int PropertyMainThirdBlendMask = Shader.PropertyToID("_Main3rdBlendMask");
-        private static readonly int PropertyUseMainSecondTex = Shader.PropertyToID("_UseMain2ndTex");
-        private static readonly int PropertyUseMainThirdTex = Shader.PropertyToID("_UseMain3rdTex");
-        private static readonly int PropertyMainColorThird = Shader.PropertyToID("_MainColor3rd");
-        private static readonly int PropertyMainThirdTexBlendMode = Shader.PropertyToID("_Main3rdTexBlendMode");
-        private static readonly int PropertyMainColorSecond = Shader.PropertyToID("_MainColor2nd");
-        private static readonly int PropertyMainSecondTexBlendMode = Shader.PropertyToID("_Main2ndTexBlendMode");
-        private static readonly int PropertyMainSecondTexAlphaMode = Shader.PropertyToID("_Main2ndTexAlphaMode");
-        private static readonly int PropertyMainThirdTexAlphaMode = Shader.PropertyToID("_Main3rdTexAlphaMode");
-        private static readonly int PropertyAlphaMask = Shader.PropertyToID("_AlphaMask");
-        private static readonly int PropertyMatCapColor = Shader.PropertyToID("_MatCapColor");
-        private static readonly int PropertyMatCapTex = Shader.PropertyToID("_MatCapTex");
-        private static readonly int PropertyUseShadow = Shader.PropertyToID("_UseShadow");
-        private static readonly int PropertyShadowColor = Shader.PropertyToID("_ShadowColor");
-        private static readonly int PropertyShadowColorTex = Shader.PropertyToID("_ShadowColorTex");
-        private static readonly int PropertyShadowStrength = Shader.PropertyToID("_ShadowStrength");
-        private static readonly int PropertyShadowStrengthMask = Shader.PropertyToID("_ShadowStrengthMask");
-        private static readonly int PropertyShadowMainStrength = Shader.PropertyToID("_ShadowMainStrength");
-        private static readonly int PropertyShadowSecondColor = Shader.PropertyToID("_Shadow2ndColor");
-        private static readonly int PropertyShadowSecondColorTex = Shader.PropertyToID("_Shadow2ndColorTex");
-        private static readonly int PropertyShadowThirdColor = Shader.PropertyToID("_Shadow3rdColor");
-        private static readonly int PropertyShadowThirdColorTex = Shader.PropertyToID("_Shadow3rdColorTex");
-        private static readonly int PropertyMainGradationTex = Shader.PropertyToID("_MainGradationTex");
-        private static readonly int PropertyMainColorAdjustMask = Shader.PropertyToID("_MainColorAdjustMask");
-        private static readonly int PropertyMainSecondTexAngle = Shader.PropertyToID("_Main2ndTexAngle");
-        private static readonly int PropertyMainSecondTexIsDecal = Shader.PropertyToID("_Main2ndTexIsDecal");
-        private static readonly int PropertyMainSecondTexIsLeftOnly = Shader.PropertyToID("_Main2ndTexIsLeftOnly");
-        private static readonly int PropertyMainSecondTexIsRightOnly = Shader.PropertyToID("_Main2ndTexIsRightOnly");
-        private static readonly int PropertyMainSecondTexShouldCopy = Shader.PropertyToID("_Main2ndTexShouldCopy");
-        private static readonly int PropertyMainSecondUVMode = Shader.PropertyToID("_Main2ndTex_UVMode");
-
-        private static readonly int PropertyMainSecondTexShouldFlipMirror =
-            Shader.PropertyToID("_Main2ndTexShouldFlipMirror");
-
-        private static readonly int PropertyMainSecondTexShouldFlipCopy =
-            Shader.PropertyToID("_Main2ndTexShouldFlipCopy");
-
-        private static readonly int PropertyMainSecondTexIsMsdf = Shader.PropertyToID("_Main2ndTexIsMSDF");
-        private static readonly int PropertyMainThirdTexAngle = Shader.PropertyToID("_Main3rdTexAngle");
-        private static readonly int PropertyMainThirdTexIsDecal = Shader.PropertyToID("_Main3rdTexIsDecal");
-        private static readonly int PropertyMainThirdTexIsLeftOnly = Shader.PropertyToID("_Main3rdTexIsLeftOnly");
-        private static readonly int PropertyMainThirdTexIsRightOnly = Shader.PropertyToID("_Main3rdTexIsRightOnly");
-        private static readonly int PropertyMainThirdTexShouldCopy = Shader.PropertyToID("_Main3rdTexShouldCopy");
-        private static readonly int PropertyMainThirdUVMode = Shader.PropertyToID("_Main3rdTex_UVMode");
-
-        private static readonly int PropertyMainThirdTexShouldFlipMirror =
-            Shader.PropertyToID("_Main3rdTexShouldFlipMirror");
-
-        private static readonly int PropertyMainThirdTexShouldFlipCopy =
-            Shader.PropertyToID("_Main3rdTexShouldFlipCopy");
-
-        private static readonly int PropertyMainThirdTexIsMsdf = Shader.PropertyToID("_Main3rdTexIsMSDF");
-        private static readonly int PropertyAlphaMaskMode = Shader.PropertyToID("_AlphaMaskMode");
-        private static readonly int PropertyAlphaMaskScale = Shader.PropertyToID("_AlphaMaskScale");
-        private static readonly int PropertyAlphaMaskValue = Shader.PropertyToID("_AlphaMaskValue");
-    }
-#endif // NVE_HAS_LILTOON
-
-    internal static class AssetPathUtils
-    {
-        private const string CloneSuffix = "(Clone)";
-
-        public static string BasePath => "Assets/NDMF VRM Exporter";
-
-        public static string GetOutputPath(GameObject gameObject)
-        {
-            return GetBasePath(gameObject, BasePath);
-        }
-
-        public static string GetTempPath(GameObject gameObject)
-        {
-            return GetBasePath(gameObject, FileUtil.GetUniqueTempPathInProject());
-        }
-
-        public static string TrimCloneSuffix(string name)
-        {
-            while (name.EndsWith(CloneSuffix, StringComparison.Ordinal))
-            {
-                name = name[..^CloneSuffix.Length];
-            }
-
-            return name;
-        }
-
-        private static string GetBasePath(GameObject gameObject, string basePath)
-        {
-            var sceneName = !string.IsNullOrEmpty(gameObject.scene.name)
-                ? StripInvalidFileNameCharacters(gameObject.scene.name)
-                : "Untitled";
-            var gameObjectName = TrimCloneSuffix(StripInvalidFileNameCharacters(gameObject.name));
-            return $"{basePath}/{sceneName}/{gameObjectName}";
-        }
-
-        private static string StripInvalidFileNameCharacters(string name)
-        {
-            return string.Join("_", name.Split(Path.GetInvalidFileNameChars()));
-        }
     }
 
     internal sealed class NdmfVrmExporter : IDisposable
@@ -4930,554 +4308,6 @@ namespace com.github.hkrn
         private readonly IReadOnlyList<MaterialVariant> _materialVariants;
         private readonly GltfMaterialExporter _materialExporter;
     }
-
-    internal sealed class NdmfVrmExporterPlugin : Plugin<NdmfVrmExporterPlugin>
-    {
-        public override string QualifiedName => NdmfVrmExporter.PackageJson.Name;
-
-        public override string DisplayName => "NDMF VRM Exporter";
-
-        protected override void Configure()
-        {
-            var platforms = new[]
-            {
-                WellKnownPlatforms.VRChatAvatar30,
-                QualifiedName
-            };
-            InPhase(BuildPhase.Transforming).OnPlatforms(platforms, seq =>
-            {
-                seq.BeforePlugin("jp.lilxyzw.lilycalinventory")
-                    .Run("Retrieve all LI CostumeChanger components to be converted to KHR_materials_variants",
-                        RetrieveAllLiCostumeChangerComponentsPass)
-                    .Then
-                    .BeforePlugin("nadena.dev.modular-avatar")
-                    .Run("Retrieve all MA reactive components to be converted to KHR_materials_variants",
-                        RetrieveAllModularAvatarReactiveComponentsPass);
-            });
-            InPhase(BuildPhase.Optimizing).OnPlatforms(platforms, seq =>
-            {
-                seq.AfterPlugin("com.anatawa12.avatar-optimizer")
-                    .AfterPlugin("nadena.dev.modular-avatar")
-                    .AfterPlugin("net.rs64.tex-trans-tool")
-                    .Run("Export VRM 1.0 file with NDMF VRM Exporter", ExportVrmFilePass);
-            });
-        }
-
-        private static void ExportVrmFilePass(BuildContext buildContext)
-        {
-            var avatarRootObject = buildContext.AvatarRootObject;
-            if (!avatarRootObject.TryGetComponent<NdmfVrmExporterComponent>(out var component))
-            {
-                Debug.LogWarning(Translator._("component.runtime.error.validation.not-attached"));
-                return;
-            }
-
-            if (!component.enabled)
-            {
-                Debug.LogWarning(Translator._("component.runtime.error.validation.not-enabled"));
-                return;
-            }
-
-            var basePath = AssetPathUtils.GetOutputPath(avatarRootObject);
-            Directory.CreateDirectory(basePath);
-            try
-            {
-                ExportVrmFile(component, buildContext, basePath, basePath);
-            }
-            catch (ValidationException e)
-            {
-                if (e.Extras != null)
-                {
-                    ErrorReport.ReportError(Translator.Instance, ErrorSeverity.NonFatal,
-                        e.Key, e.Extras);
-                }
-                else
-                {
-                    Debug.Log(e.Message);
-                }
-            }
-        }
-
-        private static void RetrieveAllModularAvatarReactiveComponentsPass(BuildContext buildContext)
-        {
-#if NVE_HAS_MODULAR_AVATAR
-            if (buildContext.AvatarRootObject.TryGetComponent<NdmfVrmExporterComponent>(
-                    out var ndmfVrmExporterComponent) &&
-                !ndmfVrmExporterComponent.enableKhrMaterialsVariants)
-            {
-                Debug.Log("KHR_materials_variants will not be exported due to be disabled in VRM Exporter Description");
-                return;
-            }
-
-            var variants = new List<MaterialVariant>();
-            var installers = buildContext.AvatarRootObject.GetComponentsInChildren<ModularAvatarMenuInstaller>();
-            foreach (var installer in installers)
-            {
-                if (!installer.enabled || !installer.TryGetComponent<ModularAvatarMenuItem>(out var menuItem))
-                {
-                    Debug.Log($"MA menu installer {installer.name} is not enabled or MA menu item is not attached");
-                    continue;
-                }
-
-                if (!menuItem.enabled)
-                {
-                    Debug.Log($"MA menu item {menuItem.name} is not enabled");
-                    continue;
-                }
-
-                if (menuItem.PortableControl.Type == PortableControlType.SubMenu)
-                {
-                    var name = menuItem.name;
-                    foreach (Transform child in menuItem.transform)
-                    {
-                        if (child.TryGetComponent<ModularAvatarMaterialSwap>(out var materialSwap))
-                        {
-                            MapMaterialSwapToVariants(buildContext, materialSwap, name, ref variants);
-                        }
-                        else if (child.TryGetComponent<ModularAvatarMaterialSetter>(out var materialSetter))
-                        {
-                            MapMaterialSetterToVariants(materialSetter, name, ref variants);
-                        }
-                    }
-                }
-                else if (menuItem.TryGetComponent<ModularAvatarMaterialSwap>(out var materialSwap))
-                {
-                    MapMaterialSwapToVariants(buildContext, materialSwap, null, ref variants);
-                }
-                else if (menuItem.TryGetComponent<ModularAvatarMaterialSetter>(out var materialSetter))
-                {
-                    MapMaterialSetterToVariants(materialSetter, null, ref variants);
-                }
-            }
-
-            buildContext.GetState<List<MaterialVariant>>().AddRange(variants);
-        }
-
-        private static void MapMaterialSwapToVariants(BuildContext buildContext, ModularAvatarMaterialSwap materialSwap,
-            string? name,
-            ref List<MaterialVariant> variants)
-        {
-            var reference = materialSwap.Root.Get(materialSwap);
-            var renderers = (reference ?? buildContext.AvatarRootObject).GetComponentsInChildren<Renderer>();
-            var sourceMappings = renderers.ToDictionary(renderer => renderer, renderer => renderer.sharedMaterials);
-            var mappings = new Dictionary<Renderer, Material[]>();
-            var variantName = materialSwap.gameObject.name;
-            if (materialSwap.gameObject.TryGetComponent<ModularAvatarMenuItem>(out var menuItem))
-            {
-                variantName = menuItem.name;
-            }
-
-            foreach (var swap in materialSwap.Swaps)
-            {
-                var swapFrom = ObjectRegistry.GetReference(swap.From);
-                foreach (var (sourceRenderer, sourceMaterials) in sourceMappings)
-                {
-                    var materialIndex = 0;
-                    foreach (var sourceMaterial in sourceMaterials)
-                    {
-                        var sourceMaterialReference = ObjectRegistry.GetReference(sourceMaterial);
-                        if (swapFrom?.Object == sourceMaterialReference?.Object)
-                        {
-                            if (mappings.TryGetValue(sourceRenderer, out var materials))
-                            {
-                                materials[materialIndex] = swap.To;
-                            }
-                            else
-                            {
-                                var newMaterials = new Material[sourceMaterials.Length];
-                                newMaterials[materialIndex] = swap.To;
-                                mappings.Add(sourceRenderer, newMaterials);
-                            }
-                        }
-
-                        materialIndex++;
-                    }
-                }
-            }
-
-            variants.Add(new MaterialVariant
-            {
-                Name = !string.IsNullOrEmpty(name) ? $"{name}/{variantName}" : variantName,
-                Mappings = mappings.Select(item => new MaterialVariantMapping
-                    { Renderer = item.Key, Materials = item.Value.ToArray(), }).ToArray(),
-            });
-        }
-
-        private static void MapMaterialSetterToVariants(ModularAvatarMaterialSetter materialSetter, string? name,
-            ref List<MaterialVariant> variants)
-        {
-            var mappings = new Dictionary<Renderer, Material[]>();
-            var variantName = materialSetter.gameObject.name;
-            if (materialSetter.gameObject.TryGetComponent<ModularAvatarMenuItem>(out var menuItem))
-            {
-                variantName = menuItem.name;
-            }
-
-            foreach (var item in materialSetter.Objects)
-            {
-                var go = item.Object.Get(materialSetter);
-                if (!go.TryGetComponent<Renderer>(out var renderer))
-                {
-                    Debug.Log($"{go} in MaterialSwitchObject has no renderer");
-                    continue;
-                }
-
-                if (mappings.TryGetValue(renderer, out var materials))
-                {
-                    if (item.MaterialIndex < materials.Length)
-                    {
-                        materials[item.MaterialIndex] = item.Material;
-                    }
-                }
-                else if (item.MaterialIndex < renderer.sharedMaterials.Length)
-                {
-                    var newMaterials = new Material[renderer.sharedMaterials.Length];
-                    newMaterials[item.MaterialIndex] = item.Material;
-                    mappings.Add(renderer, newMaterials);
-                }
-            }
-
-            variants.Add(new MaterialVariant
-            {
-                Name = !string.IsNullOrEmpty(name) ? $"{name}/{variantName}" : variantName,
-                Mappings = mappings.Select(item => new MaterialVariantMapping
-                    { Renderer = item.Key, Materials = item.Value.ToArray(), }).ToArray(),
-            });
-#else
-            Debug.Log("Modular Avatar (>= 1.13) is not installed and will be skipped");
-#endif // NVE_HAS_MODULAR_AVATAR
-        }
-
-        private static void RetrieveAllLiCostumeChangerComponentsPass(BuildContext buildContext)
-        {
-#if NVE_HAS_LILYCAL_INVENTORY
-            if (buildContext.AvatarRootObject.TryGetComponent<NdmfVrmExporterComponent>(
-                    out var ndmfVrmExporterComponent) &&
-                !ndmfVrmExporterComponent.enableKhrMaterialsVariants)
-            {
-                Debug.Log("KHR_materials_variants will not be exported due to be disabled in VRM Exporter Description");
-                return;
-            }
-
-            var costumeChangers = buildContext.AvatarRootObject.GetComponentsInChildren<CostumeChanger>();
-            var costumeChangerType = typeof(CostumeChanger);
-            var menuFolderType = typeof(MenuFolder);
-            var menuItemNameChain = new List<string>();
-            const BindingFlags bindingAttrPrivate = BindingFlags.NonPublic | BindingFlags.Instance;
-            const BindingFlags bindingAttrPublic = BindingFlags.Public | BindingFlags.Instance;
-            Type? costumeType = null;
-            Type? materialReplacerType = null;
-            Type? parametersPerMenuType = null;
-            var variants = new List<MaterialVariant>();
-            foreach (var costumeChanger in costumeChangers)
-            {
-                if (!costumeChanger.enabled)
-                {
-                    Debug.Log($"LI CostumeChanger {costumeChanger.name} is disabled and will be skipped");
-                    continue;
-                }
-
-                var baseMenuName =
-                    (string)costumeChangerType.GetField("menuName", bindingAttrPrivate)!
-                        .GetValue(costumeChanger);
-                var baseParentOverride =
-                    (MenuFolder)costumeChangerType.GetField("parentOverride", bindingAttrPrivate)!.GetValue(
-                        costumeChanger);
-                var costumes =
-                    (object[])costumeChangerType.GetField("costumes", bindingAttrPrivate)!.GetValue(
-                        costumeChanger);
-                foreach (var costume in costumes)
-                {
-                    costumeType ??= costume.GetType();
-                    var parentOverride =
-                        (MenuFolder)costumeType.GetField("parentOverride", bindingAttrPublic)!
-                            .GetValue(costume);
-                    if (!parentOverride)
-                    {
-                        parentOverride = baseParentOverride;
-                    }
-
-                    var menuIsEnabled = true;
-                    menuItemNameChain.Clear();
-                    while (parentOverride)
-                    {
-                        if (!parentOverride.enabled)
-                        {
-                            menuIsEnabled = false;
-                            break;
-                        }
-
-                        var menuItemNameInner =
-                            (string)menuFolderType.GetField("menuName", bindingAttrPrivate)!.GetValue(
-                                parentOverride);
-                        menuItemNameChain.Add(menuItemNameInner);
-                        parentOverride =
-                            (MenuFolder)menuFolderType.GetField("parentOverride", bindingAttrPrivate)!.GetValue(
-                                parentOverride);
-                    }
-
-                    if (!menuIsEnabled)
-                    {
-                        Debug.Log($"LI MenuFolder {parentOverride.name} is disabled and will be skipped");
-                        continue;
-                    }
-
-                    menuItemNameChain.Reverse();
-                    var menuItemName =
-                        (string)costumeType.GetField("menuName", bindingAttrPublic)!.GetValue(costume);
-                    var parametersPerMenu =
-                        costumeType.GetField("parametersPerMenu", bindingAttrPublic)!.GetValue(costume);
-                    parametersPerMenuType ??= parametersPerMenu.GetType();
-                    var materialReplaces =
-                        (object[])parametersPerMenuType.GetField("materialReplacers", bindingAttrPublic)!
-                            .GetValue(parametersPerMenu);
-                    var mappings = new List<MaterialVariantMapping>();
-                    foreach (var materialReplace in materialReplaces)
-                    {
-                        materialReplacerType ??= materialReplace.GetType();
-                        var fields = materialReplacerType.GetFields();
-                        var renderer = (Renderer)fields.First(item => item.Name == "renderer")
-                            .GetValue(materialReplace);
-                        var replaceTo = (Material[])fields.First(item => item.Name == "replaceTo")
-                            .GetValue(materialReplace);
-                        mappings.Add(new MaterialVariantMapping
-                        {
-                            Renderer = renderer,
-                            Materials = replaceTo,
-                        });
-                    }
-
-                    if (mappings.Count <= 0)
-                    {
-                        continue;
-                    }
-
-                    var menuItemNameChainInner = new List<string>(menuItemNameChain.AsReadOnly())
-                    {
-                        baseMenuName,
-                        menuItemName
-                    };
-                    variants.Add(new MaterialVariant
-                    {
-                        Name = string.Join("/", menuItemNameChainInner),
-                        Mappings = mappings.ToArray(),
-                    });
-                }
-            }
-
-            buildContext.GetState<List<MaterialVariant>>().AddRange(variants);
-#else
-            Debug.Log("lilycalInventory is not installed and will be skipped");
-#endif // NVE_HAS_LILYCAL_INVENTORY
-        }
-
-        internal class ValidationException : Exception
-        {
-            public string Key { get; init; } = null!;
-            public object? Extras { get; init; }
-
-            public override string Message => Translator._(Key);
-        }
-
-        internal static void ExportVrmFile(NdmfVrmExporterComponent component, BuildContext buildContext,
-            string baseOutputPath, string workingDirectoryPath)
-        {
-            if (!component.HasAuthor)
-            {
-                throw new ValidationException
-                {
-                    Key = "component.runtime.error.validation.author"
-                };
-            }
-
-            if (!component.HasLicenseUrl)
-            {
-                throw new ValidationException
-                {
-                    Key = "component.runtime.error.validation.license-url"
-                };
-            }
-
-            var corrupted = new List<SkinnedMeshRenderer>();
-            CheckAllSkinnedMeshRenderers(buildContext.AvatarRootTransform, ref corrupted);
-            if (corrupted.Count > 0)
-            {
-                throw new ValidationException
-                {
-                    Key = "component.runtime.error.validation.smr",
-                    Extras = corrupted,
-                };
-            }
-
-            var ro = buildContext.AvatarRootObject;
-            var variants = buildContext.GetState<List<MaterialVariant>>().AsReadOnly();
-            using var stream = new MemoryStream();
-            using var exporter = new NdmfVrmExporter(ro, buildContext.AssetSaver, variants);
-            var json = exporter.Export(stream);
-            var bytes = stream.ToArray();
-            var vrmFilePath = $"{baseOutputPath}.vrm";
-            NdmfVrmExporter.ReplaceFile(vrmFilePath, bytes);
-            if (component.enableGenerateJsonFile)
-            {
-                var jsonFilePath = $"{baseOutputPath}.json";
-                NdmfVrmExporter.ReplaceFile(jsonFilePath, Encoding.UTF8.GetBytes(json));
-            }
-
-            if (!component.deleteTemporaryObjects)
-            {
-                return;
-            }
-
-            foreach (var item in new[] { AssetPathUtils.GetTempPath(ro), workingDirectoryPath })
-            {
-                try
-                {
-                    var info = new DirectoryInfo(item);
-                    info.Delete(true);
-                }
-                catch (DirectoryNotFoundException)
-                {
-                    /* this is ignorable */
-                }
-            }
-        }
-
-        private static void CheckAllSkinnedMeshRenderers(Transform parent, ref List<SkinnedMeshRenderer> corrupted)
-        {
-            foreach (Transform child in parent)
-            {
-                if (!child.gameObject.activeInHierarchy)
-                {
-                    continue;
-                }
-
-                if (child.gameObject.TryGetComponent<SkinnedMeshRenderer>(out var smr) &&
-                    smr.sharedMesh && IsSharedMeshCorrupted(smr))
-                {
-                    corrupted.Add(smr);
-                }
-
-                CheckAllSkinnedMeshRenderers(child, ref corrupted);
-            }
-        }
-
-        private static bool IsSharedMeshCorrupted(SkinnedMeshRenderer smr)
-        {
-            var mesh = smr.sharedMesh;
-            var bones = smr.bones;
-            var numPositions = mesh.boneWeights.Length;
-            var numSubMeshes = mesh.subMeshCount;
-            var indexMapping = new Dictionary<uint, uint>();
-            for (var meshIndex = 0; meshIndex < numSubMeshes; meshIndex++)
-            {
-                var indices = mesh.GetIndices(meshIndex).Select(index => (uint)index).ToArray();
-                var indexSet = new HashSet<uint>(indices);
-                indexMapping.Clear();
-                for (uint i = 0; i < numPositions; i++)
-                {
-                    if (!indexSet.Contains(i))
-                    {
-                        continue;
-                    }
-
-                    var newIndex = (uint)indexMapping.Count;
-                    indexMapping.Add(i, newIndex);
-                }
-
-                if (indices.Any(index => !indexMapping.TryGetValue(index, out _)))
-                {
-                    return true;
-                }
-
-                var boneWeights = mesh.boneWeights;
-                if (boneWeights.Any(boneWeight => (boneWeight.weight0 > 0.0 && !bones[boneWeight.boneIndex0]) ||
-                                                  (boneWeight.weight1 > 0.0 && !bones[boneWeight.boneIndex1]) ||
-                                                  (boneWeight.weight2 > 0.0 && !bones[boneWeight.boneIndex2]) ||
-                                                  (boneWeight.weight3 > 0.0 && !bones[boneWeight.boneIndex3])))
-                {
-                    return true;
-                }
-            }
-
-            return false;
-        }
-    }
-
-#if NVE_HAS_AVATAR_OPTIMIZER
-    namespace aao
-    {
-        [ComponentInformation(typeof(NdmfVrmExporterComponent))]
-        internal sealed class NdmfVrmExporterComponentInformation : ComponentInformation<NdmfVrmExporterComponent>
-        {
-            protected override void CollectMutations(NdmfVrmExporterComponent component,
-                ComponentMutationsCollector collector)
-            {
-                var blendShapeNameToSmr = new Dictionary<string, SkinnedMeshRenderer>();
-                VrmExpressionProperty.RetrieveAllBlendShapes(component.transform, (name, smr, props) =>
-                {
-                    if (props.TryGetValue(name, out var foundSmr))
-                    {
-                        if (smr != foundSmr)
-                        {
-                            Debug.LogWarning($"Different SMR found: left={smr}, right={foundSmr}");
-                        }
-                    }
-                    else
-                    {
-                        props.Add(name, smr);
-                    }
-                }, ref blendShapeNameToSmr);
-                var allExpressionUsedBlendShapeNames = new List<string?>();
-                allExpressionUsedBlendShapeNames.AddRange(component.expressionPresetHappyBlendShape.BlendShapeNames);
-                allExpressionUsedBlendShapeNames.AddRange(component.expressionPresetAngryBlendShape.BlendShapeNames);
-                allExpressionUsedBlendShapeNames.AddRange(component.expressionPresetSadBlendShape.BlendShapeNames);
-                allExpressionUsedBlendShapeNames.AddRange(component.expressionPresetRelaxedBlendShape.BlendShapeNames);
-                allExpressionUsedBlendShapeNames.AddRange(component.expressionPresetSurprisedBlendShape
-                    .BlendShapeNames);
-                foreach (var property in component.expressionCustomBlendShapes)
-                {
-                    allExpressionUsedBlendShapeNames.AddRange(property.BlendShapeNames);
-                }
-
-                var collectorProperties = new Dictionary<SkinnedMeshRenderer, IList<string>>();
-                foreach (var name in allExpressionUsedBlendShapeNames)
-                {
-                    if (string.IsNullOrEmpty(name) || !blendShapeNameToSmr.TryGetValue(name!, out var smr))
-                    {
-                        continue;
-                    }
-
-                    if (collectorProperties.TryGetValue(smr, out var names))
-                    {
-                        names.Add($"blendShape.{name}");
-                    }
-                    else
-                    {
-                        collectorProperties.Add(smr, new List<string>());
-                    }
-                }
-
-                foreach (var (smr, names) in collectorProperties)
-                {
-                    collector.ModifyProperties(smr, names);
-                }
-            }
-
-            protected override void CollectDependency(NdmfVrmExporterComponent component,
-                ComponentDependencyCollector collector)
-            {
-                if (component.HasAvatarRoot && AmbientPlatform.CurrentPlatform.QualifiedName ==
-                    NdmfVrmExporterPlatform.Instance.QualifiedName)
-                {
-                    // Adds avatar's root transform as dependency to prevent removing disabled NdmfVrmExporterComponent instance
-                    collector.AddDependency(dependant: component.gameObject.transform, dependency: component);
-                }
-
-                collector.MarkEntrypoint();
-            }
-        }
-    }
-#endif // NVE_HAS_AVATAR_OPTIMIZER
 }
 
 #endif // NVE_HAS_NDMF
