+++ Editor/NdmfVrmExporterPlugin.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0549c02256c44819b4de825520144279
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
