+++ Editor/UnityExtensions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aa16eb3587fd4c99aa957909890507fb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
